requirements_20201008.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
pefile==2019.4.18
Pillow==7.2.0
playsound==1.2.2
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pyperclip==1.8.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqtkeybind==0.0.6
pywin32-ctypes==0.2.0
requests==2.24.0
urllib3==1.25.10
